v1.7

+++ src/main/java/com/ahd/api/almaktab/AlmaktabService.java
@@ -1,12 +1,16 @@
 package com.ahd.api.almaktab;
+import com.ahd.api.exception.BadRequest;
+import com.ahd.api.exception.NotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
+import org.springframework.web.client.HttpClientErrorException;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.NoSuchElementException;
 import java.util.Optional;
 
 
@@ -17,7 +21,16 @@ public class AlmaktabService {
     private AlmaktabRepository almaktabRepository;
 
     public AlmaktabModel getAlmaktab(int id) {
-        return almaktabRepository.findById(id).get();
+        try {
+            if (almaktabRepository.existsById(id)) {
+                return almaktabRepository.findById(id).get();
+            }
+        } catch (NoSuchElementException exception) {
+            throw new NotFoundException(String.format("Not Found this id [%s]", id));
+        } catch (HttpClientErrorException.BadRequest exception) {
+            throw new BadRequest("illegal request");
+        }
+        throw new NotFoundException(String.format("Not Found this id [%s]", id));
     }
 
     public List<AlmaktabModel> getListAlmaktab(Integer pageNo, Integer pageSize, String sortBy) {
@@ -26,7 +39,7 @@ public class AlmaktabService {
         if(pagedResult.hasContent()) {
             return pagedResult.getContent();
         } else {
-            return new ArrayList<AlmaktabModel>();
+            return new ArrayList<>();
         }
     }
 
@@ -35,11 +48,30 @@ public class AlmaktabService {
     }
 
     public AlmaktabModel updateAlmaktab(AlmaktabModel almaktab) {
-        return almaktabRepository.save(almaktab);
+        try {
+            if (almaktabRepository.existsById(almaktab.getId())) {
+                return almaktabRepository.save(almaktab);
+            }
+        }catch (NoSuchElementException exception) {
+            throw new NotFoundException(String.format("Not Found this id [%s]", almaktab.getId()));
+        }catch (HttpClientErrorException.BadRequest exception){
+            throw new BadRequest("illegal request");
+        }
+        throw new NotFoundException(String.format("Not Found this id [%s]", almaktab.getId()));
     }
 
-    public void deleteAlmaktab(int id) {
-        almaktabRepository.deleteById(id);
+    public boolean deleteAlmaktab(int id) {
+        try {
+            if (almaktabRepository.existsById(id)) {
+                almaktabRepository.deleteById(id);;
+                return true;
+            }
+        }catch (NoSuchElementException exception) {
+            throw new NotFoundException(String.format("Not Found this id [%s]", id));
+        }catch (HttpClientErrorException.BadRequest exception){
+            throw new BadRequest("illegal request");
+        }
+        throw new NotFoundException(String.format("Not Found this id [%s]", id));
     }
 
     public AlmaktabModel findByName(String name) {
